+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This project follows [Semantic Versioning](https://semver.org/).
 - Initial `CHANGELOG.md`
 - Planned feature list in `roadmap.md`
 - New shellscript `run-dev.sh`
-- Invalid timer feedback - if everything is wrong, it won't start and will display a error message. If at least one option is right, it will work.
+- Invalid timer feedback
 
 ### Changed
 - Improved `README.md` and docs